requirements.txt
spotipy
python-dotenv
colorama
